Ensure missing codename gracefully skips tests
